+++ warp/builtins.py
@@ -502,6 +502,37 @@ add_builtin(
     missing_grad=True,
 )
 
+add_builtin(
+    "clamp",
+    input_types={
+        "a": vector(length=Any, dtype=Scalar),
+        "low": vector(length=Any, dtype=Scalar),
+        "high": vector(length=Any, dtype=Scalar),
+    },
+    constraint=sametypes,
+    value_func=sametypes_create_value_func(vector(length=Any, dtype=Scalar)),
+    doc="Clamp the elements of ``a`` to the elements from the range [low, high].",
+    group="Vector Math",
+)
+
+add_builtin(
+    "abs",
+    input_types={"a": vector(length=Any, dtype=Scalar)},
+    constraint=sametypes,
+    value_func=sametypes_create_value_func(vector(length=Any, dtype=Scalar)),
+    doc="Return the absolute values of the elements of ``a``.",
+    group="Vector Math",
+)
+
+add_builtin(
+    "sign",
+    input_types={"a": vector(length=Any, dtype=Scalar)},
+    constraint=sametypes,
+    value_func=sametypes_create_value_func(Scalar),
+    doc="Return -1 for the negative elements of ``a``, and 1 otherwise.",
+    group="Vector Math",
+)
+
 
 def outer_value_func(arg_types: Mapping[str, type], arg_values: Mapping[str, Any]):
     if arg_types is None: