+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - Add documentation for dynamic loop autograd limitations
 - Conform to Python's syntax for function arguments when calling built-ins inside of kernels, thus extending support for keyword arguments
 - Implement the assignment operator for `wp.quat`
+- Add `wp.abs()`, `wp.clamp()`, and `wp.sign()` for vectors
 
 ## [1.2.1] - 2024-06-14